refactor: delete the unreachable inline diff, move TexraDiffView to desktop
The approval panel's inline Monaco diff never ran. `hasInlineDiff` required
originalContent/proposedContent on the ToolEditPermission payload, and
prepareToolEditApprovalPrompt — the only builder of that payload
(src/tools/approval/toolEditApproval.ts:140-150) — never set them. Both
hosts pass the payload through untouched, so the branch was dead in every
host from the day it landed.

What actually works, and is untouched here: the diff button emits openDiff,
and each host answers with its own surface — the extension opens a VS Code
diff tab (VscodeDiffViewHost), the desktop posts desktop:showDiff to its
Review workbench (desktopDiffHost). That is the whole story; the inline
panel was a third path that no one could reach.

So delete it: the inlineDiffOpen state, renderInlineDiff, hasInlineDiff and
its label/tooltip branching, and the two optional schema fields whose only
consumer it was. This supersedes #9887, which guarded that branch against a
host that could not render it — a real guard on an unreachable path.

With the last shared reference gone, TexraDiffView is desktop-only
(renderer/main.ts and reviewPane.ts), so it moves out of the extension's
component library into packages/desktop/src/renderer/, and its test moves to
test-kernel/desktop/ to match.

That move needs one new file. Per-file transpilers (esbuild under Vite and
Vitest) resolve a tsconfig by walking up from each source file: the desktop
package's tsconfig.json scopes `include` to vite.config.ts, and the repo-root
tsconfig excludes packages/desktop outright, so a Lit @customElement class
there reached the runtime with its decorator syntax untranspiled and failed
to parse. packages/desktop/src/renderer/tsconfig.json is a nearest-tsconfig
marker extending the renderer config that tsc already uses; the build scripts
still pass their configs with -p, so nothing else changes. It is the first
decorator-based component in that package, which is why nothing caught this
before.

Verified: typecheck clean; lint clean; dead-code ratchet 16 vs 16 baselined;
prettier clean; guidance references OK; progressView + desktop suites pass
(114 files, 1032 tests); the desktop renderer production build succeeds with
the component in its new home.
